+++ src/piprints/camera/__init__.py
@@ -2,6 +2,7 @@
 
 from piprints.camera.base import Camera, PreviewFrame
 from piprints.camera.exceptions import (
+    CameraCaptureError,
     CameraNotStartedError,
     CameraPreviewError,
     CameraStartupError,
@@ -10,6 +11,7 @@ from piprints.camera.picamera import PiCamera
 
 __all__ = [
     "Camera",
+    "CameraCaptureError",
     "CameraNotStartedError",
     "CameraPreviewError",
     "CameraStartupError",